Refactor: organize serializers based on models, update Swagger tags
